pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyjsclear"
dynamic = ["version"]
description = "Pure Python JavaScript deobfuscator"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = ["esprima2>=5.0.1"]
keywords = ["javascript", "deobfuscator", "deobfuscation", "security", "malware-analysis", "ast"]
authors = [
    {name = "Intezer Labs", email = "info@intezer.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/intezer/PyJSClear"
Repository = "https://github.com/intezer/PyJSClear"
Issues = "https://github.com/intezer/PyJSClear/issues"

[project.scripts]
pyjsclear = "pyjsclear.__main__:main"

[tool.setuptools.packages.find]
include = ["pyjsclear*"]

[tool.setuptools.dynamic]
version = {attr = "pyjsclear.__version__"}

[tool.black]
line-length = 120
target-version = ['py311']
skip-string-normalization = true
extend-exclude = 'tests/resources/|\.venv|\.nodeenv'

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2
line_length = 120
skip = ["tests/resources", ".venv", ".nodeenv"]

[tool.commitizen]
name = "cz_customize"

[tool.commitizen.customize]
schema_pattern = """(?s)\
(build|ci|docs|feat|fix|perf|refactor|style|test|chore|revert|bump)\
(\\(\\S+\\))?!?:\
( [^\\n\\r]+)\
((\\n\\n.*)|(\\s*))?$"""
